+++ changelog.md
@@ -6,6 +6,9 @@ This change log records updates after 2025-3-24.
 Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more info.
 
 ## [v0.6.0] - 2025-5-3
+### Added
+- Ensure `fit` validates optimizer and layers before training
+
 ### Changed
 - Refactor parameter updates to use parallel processing